+++ src/outputtypes.jl
@@ -101,15 +101,14 @@ XMetadata{T}(meta::XMetadata) where T = XMetadata{T}(
     T(meta.atol),
     (top = T(meta.lidarrange.top), bottom = T(meta.lidarrange.bottom)),
     (coarse = T.(meta.lidarprofile.coarse), fine = T.(meta.lidarprofile.fine),
-        ibottom = meta.lidarprofile.ibottom, itop = meta.lidarprofile.itop,
-        i30 = meta.lidarprofile.i30),
+        i = meta.lidarprofile.i),
     meta.sattype,
     meta.satdates,
     T(meta.altmin),
     meta.flightdates,
     meta.created,
     meta.loadtime,
-    meta.attachments
+    copy_for_promotion(meta.attachments)
 )
 
 
@@ -242,6 +241,9 @@ struct XData{T} <: Intersection{T}
         accuracy::DataFrame,
         metadata::XMetadata{T}
     ) where T
+        data = copy_for_promotion(data)
+        observations = copy_for_promotion(observations)
+        accuracy = copy_for_promotion(accuracy)
         # Ensure floats of correct precision
         convert_floats!(data, T)
         convert_floats!(accuracy, T)
@@ -360,7 +362,19 @@ function XData{T}(
 end #constructor 2 XData
 
 #* Constructor for type promotion from XData with different float precision
-XData{T}(X::XData) where T = XData{T}(X.data, X.observations, X.accuracy, XMetadata{T}(X.metadata))
+function XData{T}(X::XData) where T
+    observations = copy_for_promotion(X.observations)
+    if hasproperty(observations, :primary)
+        observations.primary = [ismissing(x) ? missing : FlightData{T}(x) for x in observations.primary]
+    end
+    if hasproperty(observations, :CPro)
+        observations.CPro = [ismissing(x) ? missing : CPro{T}(x) for x in observations.CPro]
+    end
+    if hasproperty(observations, :CLay)
+        observations.CLay = [ismissing(x) ? missing : CLay{T}(x) for x in observations.CLay]
+    end
+    XData{T}(copy_for_promotion(X.data), observations, copy_for_promotion(X.accuracy), XMetadata{T}(X.metadata))
+end
 
 #* Constructor for type promtion with different precisions for primary and secondary data
 function XData(